Cast MYSQLPORT to int for database connection

+++ php/config.php
@@ -3,9 +3,9 @@ $host = getenv('MYSQLHOST');
 $user = getenv('MYSQLUSER');
 $pass = getenv('MYSQLPASSWORD');
 $db   = getenv('MYSQLDATABASE');
-$port = getenv('MYSQLPORT');
+$port = (int)getenv('MYSQLPORT');
 
-$conn = mysqli_connect($host, $user, $pass, $db, (int)$port);
+$conn = mysqli_connect($host, $user, $pass, $db, $port);
 
 if (!$conn) {
     die("Connection failed: " . mysqli_connect_error());
